add BPF socket filter

+++ src/dp.h
@@ -6,7 +6,6 @@
 #ifndef _DP_H_
 #define _DP_H_
 
-
 #define ETH_MTU					1514
 #define MSG_LEN					4074
 
@@ -14,7 +13,7 @@
  * msg from dp sock
  *----------------------------------------*/
 typedef struct {
-	uint16_t  		port_no;
+	uint32_t  		port_no;
 	int 			sockfd;
 	char          	buffer[ETH_MTU];
 }tDP_MSG;
